Market Trend/backend/requirements.txt
fastapi>=0.103
uvicorn>=0.22
yfinance>=0.2.40
pydantic>=2.6
python-multipart>=0.0.7
sqlalchemy>=2.0
redis>=5.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
bcrypt>=4.0.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=1.40.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-exporter-prometheus>=1.12.0rc1

# Health checks and metrics
psutil>=5.9.0
aiofiles>=23.2.1

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
